requirements.txt
# CineMatch — AI Movie Finder
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
python-multipart>=0.0.6
google-generativeai>=0.8.0
httpx>=0.28.0
python-dotenv>=1.0.0
elevenlabs>=1.0.0
